requirements.txt
accelerate==0.34.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.9.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.2
datasets==2.19.2
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.0
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.7.0
entrypoints==0.4
eval_type_backport==0.2.2
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flash_attn==2.7.4.post1
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
h11==0.16.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
intel-openmp==2024.1.0
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jiter==0.9.0
jupyter-client==7.3.4
jupyter_core==5.7.2
liger_kernel_nightly==0.2.1.dev20240911164559
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2024.1.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
monai==1.4.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.77.0
opencv-python==4.11.0.86
packaging==24.0
pandas==2.2.2
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
proglog==0.1.11
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pynvml==11.5.0
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qwen-vl-utils==0.0.4
regex==2023.12.25
requests==2.32.3
rich==13.7.1
safetensors==0.4.5
sentencepiece==0.2.0
sentry-sdk==1.45.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.2
sympy==1.12
tbb==2021.12.0
tensorboardX==2.6.2.2
tokenizers==0.21.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.1
tqdm==4.66.4
traitlets==5.14.2
transformers==4.51.3
triton==2.3.1
trl==0.8.6
typing_extensions==4.13.2
tyro==0.8.4
tzdata==2024.1
ujson==5.10.0
Unidecode==1.4.0
urllib3==2.1.0
wandb==0.16.6
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
